Make Invidious instance configurable
This commit also refactors the API client methods into a class. Instances are initiated with the instance URL that should be used. This way, the API client does not have to depend on Kodi-provided libraries.

This way you can use any instance you want to, e.g., if the official one is overloaded or offline. You can even use a self-hosted Invidious instance.